# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "PPI Api",
        "version": "0.1"
    },
    "paths": {
...
        "/api/register": {
            "post": {
                "tags": [
                    "Users"
                ],
                "summary": "Register new User",
                "description": "Register new User",
                "operationId": "registerUser",
                "requestBody": {
                    "description": "User to register",
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "id": {
                                        "title": "id",
                                        "description": "id of User",
                                        "type": "integer",
                                        "readOnly": "true",
                                        "example": "1"
                                    },
                                    "first_name": {
                                        "title": "name",
                                        "description": "first name of User",
                                        "type": "string",
                                        "example": "Jozef"
                                    },
                                    "last_name": {
                                        "title": "name",
                                        "description": "first name of User",
                                        "type": "string",
                                        "example": "Bugal"
                                    },
                                    "email": {
                                        "title": "email",
                                        "description": "email of User",
                                        "type": "string"
                                    },
                                    "password": {
                                        "title": "password",
                                        "description": "password of User",
                                        "type": "password",
                                        "example": "pass12Erasd"
                                    },
                                    "password_confirmation": {
                                        "title": "password_confirmation",
                                        "description": "password confirmation of User",
                                        "type": "password",
                                        "example": "pass12Erasd"
                                    },
                                    "address": {
                                        "title": "address",
                                        "description": "address of User",
                                        "type": "string"
                                    },
                                    "date_of_birth": {
                                        "title": "date_of_birth",
                                        "description": "Users date of Birth",
                                        "type": "string",
                                        "format": "date"
                                    },
                                    "phone": {
                                        "title": "phone",
                                        "description": "Users Phone number",
                                        "type": "string"
                                    },
                                    "phone_code": {
                                        "title": "phone_code",
                                        "description": "Users Phone Code",
                                        "type": "string"
                                    },
                                    "nationality": {
                                        "title": "nationality",
                                        "description": "Users Nationality",
                                        "type": "string"
                                    },
                                    "details": {
                                        "title": "details",
                                        "description": "User Details",
                                        "type": "string"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "User registered",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "data": {
                                            "$ref": "#/components/schemas/User"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Validation exception"
                    }
                }
            }
        },
        "/api/language/store": {
            "post": {
                "tags": [
                    "Languages"
                ],
                "summary": "Create new Language",
                "description": "Create new Language",
                "operationId": "createLanguage",
                "requestBody": {
                    "description": "Language to store",
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "id": {
                                        "title": "id",
                                        "description": "id of Language",
                                        "type": "integer",
                                        "readOnly": "true",
                                        "example": "1"
                                    },
                                    "name": {
                                        "title": "name",
                                        "description": "name of Language",
                                        "type": "string",
                                        "example": "Czech"
                                    },
                                    "abbreviation": {
                                        "title": "abbreviation",
                                        "description": "abbreviation of Language",
                                        "type": "string",
                                        "example": "CZ"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Language stored",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "data": {
                                            "$ref": "#/components/schemas/Language"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Validation exception"
                    }
                }
            }
        },
        "/api/language/get": {
            "get": {
                "tags": [
                    "Languages"
                ],
                "summary": "Get all Languages",
                "description": "Get all Languages",
                "operationId": "getLanguages",
                "requestBody": {},
                "responses": {
                    "200": {
                        "description": "Get Languages",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "data": {
                                            "$ref": "#/components/schemas/Language"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Validation exception"
                    }
                }
            }
        },
...
    },
    "components": {
        "schemas": {
            "Language": {
                "title": "Language",
                "description": "Language model",
                "required": [],
                "properties": {
                    "id": {
                        "title": "id",
                        "description": "id of Language",
                        "type": "integer",
                        "readOnly": "true",
                        "example": "1"
                    },
                    "name": {
                        "title": "name",
                        "description": "name of Language",
                        "type": "string",
                        "example": "Czech"
                    },
                    "CZ": {
                        "title": "abbreviation",
                        "description": "abbreviation of Language",
                        "type": "string"
                    }
                },
                "type": "object"
            },
...
            "User": {
                "title": "User",
                "description": "User model",
                "required": [],
                "properties": {
                    "id": {
                        "title": "id",
                        "description": "id of User",
                        "type": "integer",
                        "readOnly": "true",
                        "example": "1"
                    },
                    "name": {
                        "title": "name",
                        "description": "name of User",
                        "type": "string",
                        "example": "Jozef Bugal"
                    },
                    "email": {
                        "title": "email",
                        "description": "email of User",
                        "type": "string"
                    },
                    "password": {
                        "title": "password",
                        "description": "password of User",
                        "type": "password",
                        "example": "pass12Erasd"
                    },
                    "password_confirmation": {
                        "title": "password_confirmation",
                        "description": "password confirmation of User",
                        "type": "password",
                        "example": "pass12Erasd"
                    }
                },
                "type": "object"
            },
...
        },
...
    }
}